ci(build_matrix): extract matrix_model
